hw4/requirements.txt
google-cloud-storage==2.14.0
google-cloud-pubsub==2.18.4